requirements.txt
click==8.1.3
email_validator==2.0.0.post2
Flask==2.3.2
flask_behind_proxy==0.1.1
flask_sqlalchemy==3.0.5
Flask-WTF==1.1.1
GitPython==3.1.31
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycodestyle
pytest
Werkzeug==2.3.6
newsapi-python
python-dotenv
google-generativeai